64-bit support and compiler warning fixes.

+++ HMBlkAppKit/source/HMDTAppKit/HMBlkPanel.m
@@ -99,10 +99,7 @@ POSSIBILITY OF SUCH DAMAGE.
 
 #pragma mark NSWindow
 
-- (id)initWithContentRect:(NSRect)contentRect 
-      styleMask:(unsigned int)styleMask 
-      backing:(NSBackingStoreType)backingType 
-      defer:(BOOL)flag
+- (id)initWithContentRect:(NSRect)contentRect styleMask:(NSUInteger)styleMask backing:(NSBackingStoreType)backingType defer:(BOOL)flag
 {
 	self = [super initWithContentRect:contentRect styleMask:NSBorderlessWindowMask backing:backingType defer:flag];
 	_resizable = !!(NSResizableWindowMask & styleMask);
